Crate conrod_vulkano

Structs

DrawCommand
A draw command that maps directly to the AutoCommandBufferBuilder::draw method. By returning DrawCommands, we can avoid consuming the entire AutoCommandBufferBuilder itself which might not always be available from APIs that wrap Vulkan.

GlyphCacheCommand
An command for uploading an individual glyph.

Image
A loaded vulkan texture and it's width/height

Renderer
A type used for translating render::Primitives into Commands that indicate how to draw the conrod GUI using vulkano.

Vertex
The data associated with a single vertex.

Enums

DrawError
Errors that might occur during draw calls.

RendererCreationError
Errors that might occur during creation of the renderer.